Extract gRPC methods to DemoApi
