+++ airframe-http-grpc/src/test/scala/wvlet/airframe/http/grpc/GrpcServiceBuilderTest.scala
@@ -13,180 +13,63 @@
  */
 package wvlet.airframe.http.grpc
 
-import io.grpc.stub.{AbstractBlockingStub, ClientCallStreamObserver, ClientCalls}
-import io.grpc.{CallOptions, Channel, ManagedChannel, ManagedChannelBuilder}
+import io.grpc.{Channel, ManagedChannel}
 import wvlet.airframe.Design
-import wvlet.airframe.codec.MessageCodecFactory
-import wvlet.airframe.http.router.Route
-import wvlet.airframe.http.{RPC, Router}
-import wvlet.airframe.msgpack.spi.MsgPack
-import wvlet.airframe.rx.{Rx, RxStream}
+import wvlet.airframe.http.Router
+import wvlet.airframe.http.grpc.example.DemoApi
+import wvlet.airframe.http.grpc.example.DemoApi.DemoApiClient
 import wvlet.airspec.AirSpec
+import wvlet.airframe.rx.Rx
 
 /**
   */
 object GrpcServiceBuilderTest extends AirSpec {
 
-  @RPC
-  trait MyApi {
-    def hello(name: String): String = {
-      s"Hello ${name}!"
-    }
-
-    def hello2(name: String, id: Int): String = {
-      s"Hello ${name}! (id:${id})"
-    }
-
-    def helloStreaming(name: String): RxStream[String] = {
-      Rx.sequence("Hello", "Bye").map(x => s"${x} ${name}!")
-    }
-
-    def helloClientStreaming(input: RxStream[String]): String = {
-      input.toSeq.mkString(", ")
-    }
-
-    def helloBidiStreaming(input: RxStream[String]): RxStream[String] = {
-      input.map(x => s"Hello ${x}!")
-    }
-  }
-
-  private val router = Router.add[MyApi]
+  private def router = Router.add[DemoApi]
   debug(router)
 
-  private def getRoute(name: String): Route = {
-    router.routes.find(_.methodSurface.name == name).getOrElse {
-      throw new IllegalArgumentException(s"Route is not found :${name}")
-    }
-  }
+  protected override def design = gRPC.server.withRouter(router).designWithChannel
 
-  // TODO: Generate this stub using sbt-airframe
-  class MyApiStub(
-      channel: Channel,
-      callOptions: CallOptions = CallOptions.DEFAULT,
-      codecFactory: MessageCodecFactory = MessageCodecFactory.defaultFactoryForJSON
-  ) extends AbstractBlockingStub[MyApiStub](channel, callOptions) {
-    override def build(channel: Channel, callOptions: CallOptions): MyApiStub = {
-      new MyApiStub(channel, callOptions)
-    }
-    private val codec = codecFactory.of[Map[String, Any]]
-    private val helloMethodDescriptor =
-      GrpcServiceBuilder.buildMethodDescriptor(getRoute("hello"), codecFactory)
-    private val hello2MethodDescriptor =
-      GrpcServiceBuilder.buildMethodDescriptor(getRoute("hello2"), codecFactory)
-    private val helloStreamingMethodDescriptor =
-      GrpcServiceBuilder.buildMethodDescriptor(getRoute("helloStreaming"), codecFactory)
-    private val helloClientStreamingMethodDescriptor =
-      GrpcServiceBuilder.buildMethodDescriptor(getRoute("helloClientStreaming"), codecFactory)
-    private val helloBidiStreamingMethodDescriptor =
-      GrpcServiceBuilder.buildMethodDescriptor(getRoute("helloBidiStreaming"), codecFactory)
-
-    def hello(name: String): String = {
-      val m = Map("name" -> name)
-      ClientCalls
-        .blockingUnaryCall(getChannel, helloMethodDescriptor, getCallOptions, codec.toMsgPack(m)).asInstanceOf[String]
-    }
-    def hello2(name: String, id: Int): String = {
-      val m = Map("name" -> name, "id" -> id)
-      ClientCalls
-        .blockingUnaryCall(getChannel, hello2MethodDescriptor, getCallOptions, codec.toMsgPack(m)).asInstanceOf[String]
-    }
-    def helloStreaming(name: String): Seq[String] = {
-      val m                = Map("name" -> name)
-      val responseObserver = GrpcClientCalls.blockingResponseObserver[String]
-      ClientCalls.asyncServerStreamingCall(
-        getChannel.newCall(
-          helloStreamingMethodDescriptor,
-          getCallOptions
-        ),
-        codec.toMsgPack(m),
-        responseObserver
-      )
-      responseObserver.toRx.toSeq
-    }
-    def helloClientStreaming(input: Rx[String]): String = {
-      val responseObserver = GrpcClientCalls.blockingResponseObserver[String]
-      val requestObserver: ClientCallStreamObserver[MsgPack] = ClientCalls
-        .asyncClientStreamingCall[MsgPack, Any](
-          getChannel.newCall(
-            helloClientStreamingMethodDescriptor,
-            getCallOptions
-          ),
-          responseObserver
-        ).asInstanceOf[ClientCallStreamObserver[MsgPack]]
-
-      val c = GrpcClientCalls.readClientRequestStream(input, codecFactory.of[String], requestObserver)
-      responseObserver.toRx.toSeq.head
-    }
-
-    def helloBidiStreaming(input: Rx[String]): Rx[String] = {
-      val responseObserver = GrpcClientCalls.blockingResponseObserver[String]
-      val requestObserver: ClientCallStreamObserver[MsgPack] = ClientCalls
-        .asyncBidiStreamingCall[MsgPack, Any](
-          getChannel.newCall(
-            helloBidiStreamingMethodDescriptor,
-            getCallOptions
-          ),
-          responseObserver
-        ).asInstanceOf[ClientCallStreamObserver[MsgPack]]
-
-      val c = GrpcClientCalls.readClientRequestStream(input, codecFactory.of[String], requestObserver)
-      responseObserver.toRx
-    }
-  }
-
-  test("create a standalone gRPC server") {
+  test("Create a standalone gRPC server") { server: GrpcServer =>
     gRPC.server.withRouter(router).start { server =>
       // sanity test for launching gRPC server
     }
-  }
 
-  test(
-    "create gRPC client",
-    design = {
-      wvlet.airframe.http.grpc.gRPC.server
-        .withRouter(router)
-        .design
-        .bind[ManagedChannel].toProvider { server: GrpcServer =>
-          ManagedChannelBuilder.forTarget(server.localAddress).usePlaintext().build()
+    test("Test gRPC client methods") { (channel: Channel) =>
+      val stub = new DemoApiClient(channel)
+      val N    = 10
+
+      test("unary") {
+        for (i <- 0 to N) {
+          val ret = stub.hello("world")
+          ret shouldBe "Hello world!"
         }
-        .onShutdown { channel =>
-          channel.shutdownNow()
+      }
+
+      test("n-ary") {
+        for (i <- 0 to N) {
+          val ret2 = stub.hello2("world", i)
+          ret2 shouldBe s"Hello world! (id:${i})"
         }
-    }
-  ) { (server: GrpcServer, channel: ManagedChannel) =>
-    val stub = new MyApiStub(channel)
-    val N    = 10
-
-    test("unary") {
-      for (i <- 0 to N) {
-        val ret = stub.hello("world")
-        ret shouldBe "Hello world!"
       }
-    }
 
-    test("n-ary") {
-      for (i <- 0 to N) {
-        val ret2 = stub.hello2("world", i)
-        ret2 shouldBe s"Hello world! (id:${i})"
+      test("server streaming") {
+        val streamingResults = stub.helloStreaming("RPC").toIndexedSeq
+        streamingResults shouldBe Seq("Hello RPC!", "Bye RPC!")
       }
-    }
 
-    test("server streaming") {
-      val streamingResults = stub.helloStreaming("RPC").toIndexedSeq
-      streamingResults shouldBe Seq("Hello RPC!", "Bye RPC!")
-    }
-
-    test("client streaming") {
-      for (i <- 0 to N) {
-        val result = stub.helloClientStreaming(Rx.sequence("Apple", "Banana"))
-        result shouldBe "Apple, Banana"
+      test("client streaming") {
+        for (i <- 0 to N) {
+          val result = stub.helloClientStreaming(Rx.sequence("Apple", "Banana"))
+          result shouldBe "Apple, Banana"
+        }
       }
-    }
 
-    test("bidi streaming") {
-      for (i <- 0 to N) {
-        val result = stub.helloBidiStreaming(Rx.sequence("Apple", "Banana")).toSeq
-        result shouldBe Seq("Hello Apple!", "Hello Banana!")
+      test("bidi streaming") {
+        for (i <- 0 to N) {
+          val result = stub.helloBidiStreaming(Rx.sequence("Apple", "Banana")).toSeq
+          result shouldBe Seq("Hello Apple!", "Hello Banana!")
+        }
       }
     }
   }
